Added ability to UnCamelCase words (add spaces) in Twig Templates Added dependency on forked Twig-Extensions (bgdevlab FORKED version - as original did not include un_camel_case capability). Modified the INDEX and ACTIONS templates to use new filter `un_camel_case` and existing filter `title`.
e.g. changes 'anEntityFieldDesc' becomes 'An Entity Field Desc'

Had to Override the Sensio Generator class in DoctrineCrudGenerator to add the TwigExtension awareness to the processing of the skeleton files. Looking at the logs it appears previous revision exposed the TwigEnvironment as an extention point but this no longer exists.